+++ Sekoia.io/sekoiaio/operation_center/base_get_event.py
@@ -1,4 +1,5 @@
 import time
+from typing import Callable
 from posixpath import join as urljoin
 
 import requests
@@ -26,10 +27,12 @@ class BaseGetEvents(Action):
 
         # Configure http with retry strategy
         retry_strategy = Retry(
-            total=10,
+            total=10,  # Total number of retries for all types of errors
+            status=10,  # Number of retries specifically for responses with status codes in status_forcelist
             status_forcelist=[429, 500, 502, 503, 504],
             allowed_methods=["HEAD", "GET", "OPTIONS"],
-            backoff_factor=0.2,
+            backoff_factor=1,
+            backoff_max=120,
         )
         adapter = HTTPAdapter(max_retries=retry_strategy)
         self.http_session = requests.Session()
@@ -63,22 +66,53 @@ class BaseGetEvents(Action):
 
         return response_start.json()["uuid"]
 
-    def wait_for_search_job_execution(self, event_search_job_uuid: str) -> None:
-        # wait at most 300 sec for the event search job to conclude
-        max_wait_search = 300
+    def _wait_for_search_job_step(
+        self, event_search_job_uuid: str, should_we_wait: Callable[[int], bool], action: str, timeout: int = 300
+    ) -> None:
+        """
+        Wait for a step in the search job execution
+
+        :param event_search_job_uuid: The UUID of the event search job
+        :param should_we_wait: A function that takes the current status and returns True if we should keep waiting
+        :param action: The expected action to be performed
+        :param timeout: The maximum time to wait in seconds
+        """
         start_wait = time.time()
 
+        # Initial status check
         response_get = self.http_session.get(f"{self.events_api_path}/search/jobs/{event_search_job_uuid}", timeout=20)
         response_get.raise_for_status()
 
-        while response_get.json()["status"] != 2:
+        # Wait for the condition to be met
+        while should_we_wait(response_get.json()["status"]):
+            # Wait one second before polling again
             time.sleep(1)
+
+            # Poll the job status
             response_get = self.http_session.get(
                 f"{self.events_api_path}/search/jobs/{event_search_job_uuid}", timeout=20
             )
             response_get.raise_for_status()
-            if time.time() - start_wait > max_wait_search:
-                raise TimeoutError(f"Event search job took more than {max_wait_search}s to conclude")
+
+            # If we exceed the timeout, raise an error
+            if time.time() - start_wait > timeout:
+                raise TimeoutError(f"Event search job {event_search_job_uuid} took more than {timeout}s to {action}")
+
+    def wait_for_search_job_execution(self, event_search_job_uuid: str) -> None:
+        # Wait for job to start (20 min)
+        self._wait_for_search_job_step(
+            event_search_job_uuid,
+            lambda status: status == 0,  # Wait for status to change from 0 (not started)
+            "start",
+            1200,
+        )
+        # Wait for job to complete (30 min)
+        self._wait_for_search_job_step(
+            event_search_job_uuid,
+            lambda status: status == 1,  # Wait for status to change from 1 (in progress)
+            "complete",
+            1800,
+        )
 
     def run(self, arguments: dict):
         raise NotImplementedError()